split CondCore/L1TPlugins/src/plugin.cc in smaller files
